test(e2e): extract shared login helpers to e2e/helpers.ts (#330 follow-up)
- Add fillLoginForm with hydration-safe fill logic
- Add loginAsAdmin for admin test account
- Import helpers in auth.spec.ts, profile.spec.ts, teams.spec.ts
